backend/requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-Cors==3.0.10
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mutagen==1.47.0
pycryptodomex==3.20.0
python-dotenv==1.0.0
pytube==15.0.0
requests==2.32.3
six==1.16.0
urllib3==2.2.2
websockets==13.0.1
Werkzeug==2.3.6
youtube-dl==2021.12.17
yt-dlp==2024.8.6
